SellHouse/requirements.txt
Django==2.2.5
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
Pillow==6.1.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
PyJWT==1.7.1
pytz==2019.2
sqlparse==0.3.0
